Removed various warnings issued by Eclipse.
git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1607 267469b6-68fb-4578-8a4b-5844dc4013a7

+++ source/net/sourceforge/toscanaj/controller/fca/AbstractConceptInterperter.java
@@ -277,7 +277,7 @@ public abstract class AbstractConceptInterperter implements ConceptInterpreter, 
                         NumberFormat format = DecimalFormat.getPercentInstance();
                         format.setMaximumFractionDigits(2);
                         String returnValue = format.format(objectCount/(double)fullExtent);
-                        String expectedValue = "[exp: " + format.format(expectedSize/(double)fullExtent) + "]";
+                        String expectedValue = "[exp: " + format.format(expectedSize/fullExtent) + "]";
                         return new Object[]{getObject(returnValue, concept, context),
                                             getObject(expectedValue, concept, context)};              
                     } // else fall back into normal behaviour
